collection-based default resource determination
This commit allows the resource where a new data object's replica will
be stored to be based on the collection where the data object is
uploaded. If there is a ipc::hosted-collection AVU whose value is a
collection the data object belongs to, the replica will be stored in the
resource where the AVU is attached. If the unit is 'forced', this
resource choice cannot be overridden by the operation creating the data
object, otherwise it can.